Move datetime tests into separate file
